BIRC: Fix 64-bit compatibility
Files: cb_1/lib/cgc_stdint.h cb_1/src/cgc_server_backend.h cb_1/src/server.cc cb_1/src/server_backend.cc cb_2/lib/cgc_stdint.h cb_2/src/cgc_client_backend.h cb_2/src/client.cc cb_2/src/client_backend.cc

+++ challenges/BIRC/cb_1/src/cgc_server_backend.h
@@ -22,11 +22,12 @@
  */
 #ifndef SERVER_BACKEND_H_
 #define SERVER_BACKEND_H_
+#include <cstddef>
 #include "cgc_user.h"
 
 namespace Backend {
 
-void InitDataStores(int secret_page_i);
+void InitDataStores(size_t secret_page_i);
 
 char *AddUser(const char *name, unsigned short length, unsigned short *pass_length);
 bool CheckIfUserExists(const char *name, unsigned short length);
